+++ app/controllers/sessions_controller.rb
@@ -112,7 +112,7 @@ class SessionsController < ApplicationController
 
   def update
     if otp_valid? && authorize_guest!
-      expire_at = client.created_at < 5.minute.ago ? 24.hours.from_now : 10.years.from_now
+      expire_at = @device.client.created_at < 5.minute.ago ? 24.hours.from_now : 10.years.from_now
       Device.find(session[:did]).update!(
         last_authenticated_at: Time.current,
         authentication_expire_at: expire_at,
@@ -142,10 +142,10 @@ class SessionsController < ApplicationController
     end
 
     def find_or_create_user_client(user)
-      client = Client.where(tenant_id: user[:tid]).or(Client.where(email: user[:email])).or(Client.where(phone: user[:phone])).first_or_initialize
-      client.update!(name: user[:name], active: true) if client.name != user[:name] && user[:name].present? && user[:name] != ""
-      expire_at = client.created_at < 5.minute.ago ? 24.hours.from_now : 10.years.from_now
-      device = Device.find_or_create_by!(client_id: client.id, mac_address: user[:id])
+      @client = Client.where(tenant_id: user[:tid]).or(Client.where(email: user[:email])).or(Client.where(phone: user[:phone])).first_or_initialize
+      @client.update!(name: user[:name], active: true) if @client.name != user[:name] && user[:name].present? && user[:name] != ""
+      expire_at = @client.created_at < 5.minute.ago ? 24.hours.from_now : 10.years.from_now
+      device = Device.find_or_create_by!(client_id: @client.id, mac_address: user[:id])
       device.update!(
         last_ap: user[:ap],
         site_id: user[:sid],