+++ test/controllers/admin/sessions_controller_test.rb
@@ -2,17 +2,17 @@ require "test_helper"
 
 class Admin::SessionsControllerTest < ActionDispatch::IntegrationTest
   test "should get new" do
-    get admin_sessions_new_url
+    get admin_login_url
     assert_response :success
   end
 
   test "should get create" do
-    get admin_sessions_create_url
+    get admin_login_url
     assert_response :success
   end
 
   test "should get destroy" do
-    get admin_sessions_destroy_url
-    assert_response :success
+    delete admin_logout_url
+    assert_redirected_to admin_login_url
   end
 end