+++ test/controllers/admin/tenants_controller_test.rb
@@ -3,6 +3,8 @@ require "test_helper"
 class Admin::TenantsControllerTest < ActionDispatch::IntegrationTest
   setup do
     @admin_tenant = tenants(:one)
+    @user = users(:one)
+    post admin_login_path, params: { email: @user.email, password: "password" }
   end
 
   test "should get index" do
@@ -17,7 +19,7 @@ class Admin::TenantsControllerTest < ActionDispatch::IntegrationTest
 
   test "should create admin_tenant" do
     assert_difference("Tenant.count") do
-      post admin_tenants_url, params: { tenant: { active: @admin_tenant.active, login: @admin_tenant.login, name: @admin_tenant.name, password: @admin_tenant.password } }
+      post admin_tenants_url, params: { tenant: { active: @admin_tenant.active, name: @admin_tenant.name  } }
     end
 
     assert_redirected_to admin_tenants_url
@@ -34,7 +36,7 @@ class Admin::TenantsControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should update admin_tenant" do
-    patch admin_tenant_url(@admin_tenant), params: { tenant: { active: @admin_tenant.active, login: @admin_tenant.login, name: @admin_tenant.name, password: @admin_tenant.password } }
+    patch admin_tenant_url(@admin_tenant), params: { tenant: { active: @admin_tenant.active, name: @admin_tenant.name } }
     assert_redirected_to admin_tenants_url
   end