+++ test/controllers/clients_controller_test.rb
@@ -3,6 +3,11 @@ require "test_helper"
 class ClientsControllerTest < ActionDispatch::IntegrationTest
   setup do
     @client = clients(:one)
+    @user = users(:one)
+    post admin_login_path, params: { email: @user.email, password: "password" }
+    stub_unifi_sites_api(@client.devices.first.site.unifi_id)
+    stub_unifi_client_api(@client.devices.first.mac_address, @client.devices.first.site.unifi_id)
+    stub_guest_authorization_api("https://heimdall.test", site_id: @client.devices.first.site.unifi_id)
   end
 
   test "should get index" do