+++ app/javascript/controllers/session_controller.js
@@ -42,11 +42,12 @@ export default class extends Controller {
         .then(response => response.json())
         .then(data => {
           if (!data.exists) {
-            phoneInput.parentElement.classList.remove("border", "border-green-500");
+            phoneInput.parentElement.classList.remove("border-green-500");
             phoneInput.parentElement.classList.add("border", "border-yellow-500");
           } else {
-            phoneInput.parentElement.classList.remove("border", "border-yellow-500");
-            phoneInput.parentElement.classList.add("border", "border-green-500");
+            console.log(phoneInput);
+            phoneInput.parentElement.classList.remove("border-yellow-500");
+            phoneInput.parentElement.classList.add("border-green-500");
             document.querySelector("#name").disabled = true;
             document.querySelector("#email").disabled = true;
             document.getElementById("submit-otp").click();