+++ app/models/device.rb
@@ -22,37 +22,45 @@ class Device < ApplicationRecord
     if site.nil? || !client.active?
       return false
     end
-    load_client_info
-    result = External::Unifi.authorize_guest(
-      url: site.controller_url,
-      site_id: site.unifi_id,
-      client_id: unifi_id,
-      api_key: site.api_key,
-      time: time_limit,
-      guest_max: client.guest_max,
-      guest_rx: client.guest_rx,
-      guest_tx: client.guest_tx
-    )
-    !result.nil? && result.dig("action").present? && result["action"] == "AUTHORIZE_GUEST_ACCESS" ?
-      { success: true } :
-      { success: false, error: result["error"] || "Failed to authorize guest access" }
+    eu = External::Unifi::Base.new(site: site)
+    if eu
+      load_client_info(eu)
+      result = eu.authorize_guest_access(
+        mac_address: mac_address,
+        minutes: time_limit,
+        up: client.guest_tx,
+        down: client.guest_rx,
+        megabytes: client.guest_max
+      )
+      # result = eu.authorize_guest(
+      #   url: site.controller_url,
+      #   site_id: site.unifi_id,
+      #   client_id: unifi_id,
+      #   api_key: site.api_key,
+      #   time: time_limit,
+      #   guest_max: client.guest_max,
+      #   guest_rx: client.guest_rx,
+      #   guest_tx: client.guest_tx
+      # )
+      result ?
+        { success: true } :
+        { success: false, error: "Failed to authorize guest access" }
+    end
   end
 
   def time_limit
     authentication_expire_at > 24.hours.from_now ? 1000000 : 1440
   end
 
-  def load_client_info
-    site_info = External::Unifi.get_sites(site&.controller_url, key: site&.api_key)
-    site.update(unifi_id: site_info["data"].first["id"])
-    @unifi_client = External::Unifi.get_client(site.controller_url, site.unifi_id, mac_address, key: site&.api_key)
-    unless @unifi_client.nil? ||
-      @unifi_client.dig("count").zero? ||
-      @unifi_client.dig("data").nil? ||
-      @unifi_client.dig("data").empty? ||
-      @unifi_client.dig("count") > 1
-
-      update unifi_id: @unifi_client["data"].first["id"]
+  def load_client_info(eu)
+    site_info = eu.site_info
+    site.update(unifi_id: eu.get_id) if site_info
+    unifi_client_id = eu.get_client_id(mac_address)
+    unless unifi_client_id.nil?
+      update unifi_id: unifi_client_id
+    else
+      Rails.logger.error("ERROR: Unifi client info not found for MAC address: #{mac_address}")
+      nil
     end
   end
 end